# docs/openapi.yaml
openapi: 3.1.0
info:
  title: careerAPI
  version: 0.1.0
  description: |
    Backend for the careerAI frontend: student accounts, school-based access and
    payments, the weekly aptitude contest, and the admin panel.

    **Design:** `docs/database-schema.md` (tables, rules, access check).
    **Contest rules:** `careerAI/docs/aptitude-contest-backend.md`.
    **Frontend contract:** every endpoint here has a matching method in the
    frontend mocks (`careerAI/scripts/careerapi_mock.js`, `contest_api.js`), and
    the response shapes are the same.

    ### Sessions
    - Students log in at `/auth/*` and get the `careerai_session` cookie.
    - Admins log in at `/admin/auth/*` and get a **separate** `careerai_admin_session` cookie.
    - Both are opaque, `HttpOnly; Secure; SameSite=None; Path=/` cookies (the frontend is another site,
      or reaches the API through an `/api` proxy). The database stores only their SHA-256 hash.
    - Student sessions last 30 days and slide with use; admin sessions last 12 hours.

    ### Access (all or nothing)
    Every endpoint marked **Paid** first runs the access check (schema §6.4) and
    answers `402 payment_required` with a `reason` when the student has no access.
    A student has access when they have an active individual plan, **or** their
    School ID's school has an active plan and their email is on that school's list.

    ### Conventions
    - JSON throughout. Times are ISO-8601 UTC strings; the frontend shows them in IST.
...
servers:
...
security:
  - studentSession: []
paths:
...
  /contest/state:
    get:
      tags: [Contest]
      summary: Starting point of the contest page
      description: '**Paid.** `contest` is this week''s while it''s open, otherwise the next one.'
      responses:
        '200':
          description: State
          content:
            application/json:
              schema:
                type: object
                properties:
                  server_time: { type: string, format: date-time }
                  contest: { $ref: '#/components/schemas/Contest' }
                  attempt:
                    type: object
                    properties:
                      status: { type: string, enum: [not_started, in_progress, submitted] }
                      started_at: { type: [string, 'null'], format: date-time }
                      submitted_at: { type: [string, 'null'], format: date-time }
                      answered_count: { type: integer }
                      auto_submitted: { type: boolean }
                  latest_published: { oneOf: [{ $ref: '#/components/schemas/Contest' }, { type: 'null' }] }
                  published_count: { type: integer }
                  career: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
  /contest/{id}/attempt:
    parameters:
      - $ref: '#/components/parameters/ContestId'
    post:
      tags: [Contest]
      summary: Start (or resume) my attempt
      description: |
        **Paid.** Starts the clock on the server. Calling it again resumes the same
        attempt without resetting the clock. Questions never include the answer.
      responses:
        '200':
          description: The attempt
          content:
            application/json:
              schema:
                type: object
                properties:
                  contest: { $ref: '#/components/schemas/Contest' }
                  started_at: { type: string, format: date-time }
                  questions:
                    type: array
                    items: { $ref: '#/components/schemas/PublicQuestion' }
                  answers:
                    type: object
                    additionalProperties: { type: integer, minimum: 0, maximum: 3 }
                    example: { c5-q1: 2 }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { description: '`not_open`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: '`already_submitted`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /contest/{id}/attempt/answers/{questionId}:
    parameters:
      - $ref: '#/components/parameters/ContestId'
      - in: path
        name: questionId
        required: true
        schema: { type: string, example: c5-q1 }
    put:
      tags: [Contest]
      summary: Save my answer to one question
      description: '**Paid.** Upsert; the latest choice wins.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [option_index]
              properties:
                option_index: { type: integer, minimum: 0, maximum: 3 }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { description: '`bad_question` (not in this contest)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { description: '`not_open`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: '`already_submitted` or `not_started`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /contest/{id}/attempt/submit:
    parameters:
      - $ref: '#/components/parameters/ContestId'
    post:
      tags: [Contest]
      summary: Submit my attempt
      description: '**Paid.** Final and idempotent. The response deliberately contains **no score**; scores appear after close.'
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  submitted_at: { type: string, format: date-time }
                  time_taken_s: { type: integer, minimum: 1 }
                  results_at: { type: string, format: date-time, description: The contest's closes_at }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { description: '`not_open`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: '`not_started`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /contest/{id}/review:
    parameters:
      - $ref: '#/components/parameters/ContestId'
    get:
      tags: [Contest]
      summary: Answers and my result, after close
      description: '**Paid.** Available to every student after close, including those who didn''t take part.'
      responses:
        '200':
          description: Review
          content:
            application/json:
              schema:
                type: object
                properties:
                  contest: { $ref: '#/components/schemas/Contest' }
                  attempted: { type: boolean }
                  score: { type: [integer, 'null'] }
                  correct: { type: [integer, 'null'] }
                  total_questions: { type: integer, const: 5 }
                  time_taken_s: { type: [integer, 'null'] }
                  global: { oneOf: [{ $ref: '#/components/schemas/Rank' }, { type: 'null' }] }
                  career:
                    oneOf:
                      - allOf:
                          - $ref: '#/components/schemas/Rank'
                          - type: object
                            properties:
                              name: { type: string, example: Architect }
                      - type: 'null'
                  questions:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/PublicQuestion'
                        - type: object
                          properties:
                            correct_index: { type: integer, minimum: 0, maximum: 3 }
                            explanation: { type: string }
                            your_index: { type: [integer, 'null'] }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { description: '`not_published` (still open or upcoming)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /leaderboard:
    get:
      tags: [Contest]
      summary: Global or career leaderboard, for one week or all time
      description: '**Paid.** `me` is always included, even outside the top `limit`.'
      parameters:
        - { in: query, name: scope, schema: { type: string, enum: [global, career], default: global } }
        - { in: query, name: period, schema: { type: string, enum: [week, live], default: week } }
        - { in: query, name: contest_id, description: 'Closed contest (week only). Default: latest closed.', schema: { type: integer } }
        - { in: query, name: limit, schema: { type: integer, default: 10, maximum: 100 } }
      responses:
        '200':
          description: Board
          content:
            application/json:
              schema:
                type: object
                properties:
                  scope: { type: string, enum: [global, career] }
                  period: { type: string, enum: [week, live] }
                  career: { type: [string, 'null'], description: Career scope only }
                  contest: { oneOf: [{ $ref: '#/components/schemas/Contest' }, { type: 'null' }], description: Week only }
                  contests_count: { type: integer, description: 'Live: closed contests included' }
                  total_participants: { type: integer }
                  me: { oneOf: [{ $ref: '#/components/schemas/LeaderboardRow' }, { type: 'null' }] }
                  rows: { type: array, items: { $ref: '#/components/schemas/LeaderboardRow' } }
                  locked: { type: string, enum: [no_career], description: Only when scope=career and the student has no career yet }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { description: '`not_published`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ══════════════════════════════════════════════════════════════
  # ADMIN AUTH
  # ══════════════════════════════════════════════════════════════
...
  /admin/students:
    get:
      tags: [Admin · Students]
      summary: List students
      security: [{ adminSession: [] }]
      parameters:
        - $ref: '#/components/parameters/Deleted'
        - { in: query, name: q, description: Search name or email, schema: { type: string } }
      responses:
        '200':
          description: Students, A–Z, each with their access result
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Student'
                    - type: object
                      properties:
                        is_deleted: { type: boolean }
                        deleted_at: { type: [string, 'null'], format: date-time }
                        access: { $ref: '#/components/schemas/Access' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /admin/students/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Admin · Students]
      summary: Delete a student
      description: |
        Soft: can't log in, hidden from leaderboards, restorable.
        Hard: removes the account, profile, quiz results, attempts and answers.
        Their payment records are kept with `student_id = NULL`.
      security: [{ adminSession: [] }]
      parameters:
        - $ref: '#/components/parameters/DeleteMode'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/students/{id}/restore:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Admin · Students]
      summary: Restore a soft-deleted student
      security: [{ adminSession: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/students/{id}/password:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Admin · Students]
      summary: Reset a student's password
      description: |
        Sets a new password and signs the student out on every device.
        A soft-deleted student still can't log in until restored.
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password, minLength: 8, maxLength: 200 }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Invalid' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: '`no_password_login`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/admins:
    get:
      tags: [Admin · Admins]
      summary: List admins
      security: [{ adminSession: [] }]
      parameters:
        - $ref: '#/components/parameters/Deleted'
      responses:
        '200':
          description: Admins, A–Z
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Admin' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [Admin · Admins]
      summary: Add an admin
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, email, password]
              properties:
                full_name: { type: string }
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/Invalid' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { description: '`email_taken`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/admins/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Admin · Admins]
      summary: Delete an admin (not yourself)
      description: Hard delete keeps their contests and schools, with `created_by = NULL`.
      security: [{ adminSession: [] }]
      parameters:
        - $ref: '#/components/parameters/DeleteMode'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: '`cannot_delete_self`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/admins/{id}/restore:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Admin · Admins]
      summary: Restore a soft-deleted admin
      security: [{ adminSession: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

# ══════════════════════════════════════════════════════════════
components:
  securitySchemes:
    studentSession:
      type: apiKey
      in: cookie
      name: careerai_session
    adminSession:
      type: apiKey
      in: cookie
      name: careerai_admin_session
...
  parameters:
    ContestId:
      in: path
      name: id
      required: true
      description: Contest id, which is also its public number (#1, #2…)
      schema: { type: integer, minimum: 1 }
...
    Deleted:
      in: query
      name: deleted
      description: '`true` lists only soft-deleted rows'
      schema: { type: boolean, default: false }
    DeleteMode:
      in: query
      name: mode
      description: '`soft` sets is_deleted and can be undone. `hard` is permanent (schema §7).'
      schema: { type: string, enum: [soft, hard], default: soft }
  responses:
    Ok:
      description: Done
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, const: true }
    Invalid:
      description: Validation failed. `field` names the input to highlight.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: invalid, message: Password must be at least 8 characters., field: password } }
    Unauthenticated:
      description: '`unauthenticated`: no valid session'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthenticated, message: Please log in. } }
    PaymentRequired:
      description: '`payment_required`: the student has no access. `reason` says why, for the paywall.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, const: payment_required }
                  message: { type: string }
                  reason: { $ref: '#/components/schemas/AccessReason' }
          example: { error: { code: payment_required, message: Your school isn't covering you yet., reason: not_on_roster } }
    NotFound:
      description: '`not_found`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: invalid }
            message: { type: string, description: Safe to show to the user }
            field: { type: string, description: The request field that caused it, when there is one }
    SchoolRef:
      type: object
      properties:
        name: { type: string, example: 'Delhi Public School, R.K. Puram' }
        school_code: { type: string, example: DPS-RKP }
    Student:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string, example: Ananya Sharma }
        leaderboard_name: { type: string, example: Ananya S., description: First name + last initial (schema §2.1) }
        email: { type: string, format: email, description: Only ever returned to the student themself or to admins }
        school: { oneOf: [{ $ref: '#/components/schemas/SchoolRef' }, { type: 'null' }] }
        current_career: { type: [string, 'null'], example: UX Designer }
        created_at: { type: string, format: date-time }
    AccessReason:
      type: [string, 'null']
      enum: [no_school, not_on_roster, school_not_subscribed, null]
      description: |
        - `no_school`: no School ID given
        - `not_on_roster`: gave a School ID, but their email isn't on that school's list
        - `school_not_subscribed`: on the list, but the school has no active plan
    Access:
      type: object
      properties:
        has_access: { type: boolean }
        source: { type: [string, 'null'], enum: [school, individual, null] }
        reason: { $ref: '#/components/schemas/AccessReason' }
        school: { oneOf: [{ $ref: '#/components/schemas/SchoolRef' }, { type: 'null' }] }
        ends_at: { type: [string, 'null'], format: date-time, description: When the covering plan ends }
        plan: { type: [string, 'null'], example: school_annual }
...
    Contest:
      description: As seen by students (contest spec §4)
      type: object
      properties:
        id: { type: integer, example: 5 }
        number: { type: integer, example: 5 }
        opens_at: { type: string, format: date-time, example: '2026-10-03T01:30:00.000Z' }
        closes_at: { type: string, format: date-time, example: '2026-10-04T13:30:00.000Z' }
        status: { type: string, enum: [upcoming, open, closed] }
        question_count: { type: integer, const: 5 }
        points_per_question: { type: integer, const: 10 }
    PublicQuestion:
      description: A question without its answer
      type: object
      properties:
        id: { type: string, example: c5-q1 }
        area: { type: string, enum: [Logical Reasoning, Numerical Ability, Verbal Ability, Spatial Reasoning] }
        text: { type: string }
        options: { type: array, items: { type: string }, minItems: 4, maxItems: 4 }
    Rank:
      type: object
      properties:
        rank: { type: integer, example: 50 }
        total: { type: integer, example: 500 }
    LeaderboardRow:
      type: object
      properties:
        rank: { type: integer }
        name: { type: string, example: Ananya S. }
        career: { type: [string, 'null'], example: Architect }
        career_emoji: { type: string, example: 🏛️ }
        score: { type: integer }
        correct: { type: integer }
        total_questions: { type: integer, description: 'Week: 5. Live: 5 × contests.' }
        time_taken_s: { type: integer }
        contests: { type: integer, description: Live only }
        is_me: { type: boolean }
    AdminRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
    Admin:
      allOf:
        - $ref: '#/components/schemas/AdminRef'
        - type: object
          properties:
            last_login_at: { type: [string, 'null'], format: date-time }
            is_deleted: { type: boolean }
            deleted_at: { type: [string, 'null'], format: date-time }
            created_at: { type: string, format: date-time }
            is_me: { type: boolean }
...